Switched to double precision for nujel floats
Started writing testcases and realized just how
inaccurate floats are, vec's will still use floats
though.

+++ lib/common.h
@@ -79,7 +79,7 @@ struct lVal {
 		bool           vBool;
 		lPair          vList;
 		i64            vInt;
-		float          vFloat;
+		double         vFloat;
 		vec            vVec;
 		lArray        *vArray;
 		lTree         *vTree;
