+++ lib/operation/arithmetic.c
@@ -38,8 +38,8 @@ static vec lnfAddV(const lVal *v){
 	}
 	return acc;
 }
-static float lnfAddF(const lVal *v){
-	float acc = v->vList.car->vFloat;
+static double lnfAddF(const lVal *v){
+	double acc = v->vList.car->vFloat;
 	v = v->vList.cdr;
 	for(; v ; v = v->vList.cdr){
 		acc += v->vList.car->vFloat;
@@ -76,8 +76,8 @@ static vec lnfSubV(lVal *v){
 	}
 	return acc;
 }
-static float lnfSubF(lVal *v){
-	float acc = v->vList.car->vFloat;
+static double lnfSubF(lVal *v){
+	double acc = v->vList.car->vFloat;
 	v = v->vList.cdr;
 	if(!v){return -acc;}
 	for(; v ; v = v->vList.cdr){
@@ -113,8 +113,8 @@ static vec lnfMulV(lVal *v){
 	}
 	return acc;
 }
-static float lnfMulF(lVal *v){
-	float acc;
+static double lnfMulF(lVal *v){
+	double acc;
 	for(acc = 1.f; v ; v = v->vList.cdr){
 		acc *= v->vList.car->vFloat;
 	}
@@ -148,8 +148,8 @@ static vec lnfDivV(lVal *v){
 	}
 	return acc;
 }
-static float lnfDivF(lVal *v){
-	float acc = v->vList.car->vFloat;
+static double lnfDivF(lVal *v){
+	double acc = v->vList.car->vFloat;
 	v = v->vList.cdr;
 	for(; v ; v = v->vList.cdr){
 		acc /= v->vList.car->vFloat;
@@ -192,8 +192,8 @@ static vec lnfModV(lVal *v){
 	}
 	return acc;
 }
-static float lnfModF(lVal *v){
-	float acc = v->vList.car->vFloat;
+static double lnfModF(lVal *v){
+	double acc = v->vList.car->vFloat;
 	v = v->vList.cdr;
 	for(; v ; v = v->vList.cdr){
 		acc = fmodf(acc,v->vList.car->vFloat);
@@ -230,7 +230,7 @@ lVal *lnfAbs(lClosure *c, lVal *v){
 	if(t == NULL){return lValInt(0);}
 	switch(t->type){
 		default:      return exceptionThrow(c, v,"absolute");
-		case ltFloat: return lValFloat(fabsf(t->vFloat));
+		case ltFloat: return lValFloat(fabs(t->vFloat));
 		case ltInt:   return lValInt(llabs(t->vInt));
 		case ltVec:   return lValVec(vecAbs(t->vVec));
 	}
@@ -241,8 +241,8 @@ lVal *lnfCbrt(lClosure *c, lVal *v){
 	if(t == NULL){return lValInt(0);}
 	switch(t->type){
 		default:      return exceptionThrow(c, v,"squareroot");
-		case ltFloat: return lValFloat(cbrtf(t->vFloat));
-		case ltInt:   return lValFloat(cbrtf(t->vInt));
+		case ltFloat: return lValFloat(cbrt(t->vFloat));
+		case ltInt:   return lValFloat(cbrt(t->vInt));
 		case ltVec:   return lValVec(vecCbrt(t->vVec));
 	}
 }
@@ -252,8 +252,8 @@ lVal *lnfSqrt(lClosure *c, lVal *v){
 	if(t == NULL){return lValInt(0);}
 	switch(t->type){
 		default:      return exceptionThrow(c, v,"squareroot");
-		case ltFloat: return lValFloat(sqrtf(t->vFloat));
-		case ltInt:   return lValFloat(sqrtf(t->vInt));
+		case ltFloat: return lValFloat(sqrt(t->vFloat));
+		case ltInt:   return lValFloat(sqrt(t->vInt));
 		case ltVec:   return lValVec(vecSqrt(t->vVec));
 	}
 }
@@ -263,7 +263,7 @@ lVal *lnfCeil(lClosure *c, lVal *v){
 	if(t == NULL){return lValFloat(0);}
 	switch(t->type){
 		default:      return exceptionThrow(c, v,"ceil");
-		case ltFloat: return lValFloat(ceilf(t->vFloat));
+		case ltFloat: return lValFloat(ceil(t->vFloat));
 		case ltVec:   return lValVec(vecCeil(t->vVec));
 	}
 }
@@ -273,7 +273,7 @@ lVal *lnfFloor(lClosure *c, lVal *v){
 	if(t == NULL){return lValFloat(0);}
 	switch(t->type){
 		default:      return exceptionThrow(c, v,"floor");
-		case ltFloat: return lValFloat(floorf(t->vFloat));
+		case ltFloat: return lValFloat(floor(t->vFloat));
 		case ltVec:   return lValVec(vecFloor(t->vVec));
 	}
 }
@@ -283,7 +283,7 @@ lVal *lnfRound(lClosure *c, lVal *v){
 	if(t == NULL){return lValFloat(0);}
 	switch(t->type){
 		default:      return exceptionThrow(c, v,"round");
-		case ltFloat: return lValFloat(roundf(t->vFloat));
+		case ltFloat: return lValFloat(round(t->vFloat));
 		case ltVec:   return lValVec(vecRound(t->vVec));
 	}
 }
@@ -293,7 +293,7 @@ lVal *lnfSin(lClosure *c, lVal *v){
 	if(t == NULL){return lValFloat(0);}
 	switch(t->type){
 		default:      return exceptionThrowFloat(c, v,"sinus");
-		case ltFloat: return lValFloat(sinf(t->vFloat));
+		case ltFloat: return lValFloat(sin(t->vFloat));
 	}
 }
 
@@ -302,7 +302,7 @@ lVal *lnfCos(lClosure *c, lVal *v){
 	if(t == NULL){return lValFloat(0);}
 	switch(t->type){
 		default:      return exceptionThrowFloat(c, v,"cosine");
-		case ltFloat: return lValFloat(cosf(t->vFloat));
+		case ltFloat: return lValFloat(cos(t->vFloat));
 	}
 }
 
@@ -311,7 +311,7 @@ lVal *lnfTan(lClosure *c, lVal *v){
 	if(t == NULL){return lValFloat(0);}
 	switch(t->type){
 		default:      return exceptionThrowFloat(c, v,"tangent");
-		case ltFloat: return lValFloat(tanf(t->vFloat));
+		case ltFloat: return lValFloat(tan(t->vFloat));
 	}
 }
 
@@ -326,8 +326,8 @@ lVal *lnfPow(lClosure *c, lVal *v){
 	if(u == NULL){return lValInt(0);}
 	switch(t->type){
 		default:      return exceptionThrowFloat(c, v,"power");
-		case ltFloat: return lValFloat(powf(t->vFloat,u->vFloat));
-		case ltInt:   return lValFloat(powf(t->vInt,u->vInt));
+		case ltFloat: return lValFloat(pow(t->vFloat,u->vFloat));
+		case ltInt:   return lValFloat(pow(t->vInt,u->vInt));
 		case ltVec:   return lValVec(vecPow(t->vVec,u->vVec));
 	}
 }
